app/requirements.txt
# Add the required packages
streamlit==1.18.1
tensorflow==2.8.0
numpy==1.21.4
pandas==1.3.5
matplotlib==3.5.1
seaborn==0.11.2
Pillow==9.0.0
opencv-python==4.5.5.62

